move enterprise/internal/notebooks to internal/notebooks
