+++ enterprise/cmd/frontend/internal/notebooks/resolvers/BUILD.bazel
@@ -14,10 +14,10 @@ go_library(
         "//cmd/frontend/envvar",
         "//cmd/frontend/graphqlbackend",
         "//cmd/frontend/graphqlbackend/graphqlutil",
-        "//enterprise/internal/notebooks",
         "//internal/database",
         "//internal/errcode",
         "//internal/gqlutil",
+        "//internal/notebooks",
         "//lib/errors",
         "@com_github_graph_gophers_graphql_go//:graphql-go",
         "@com_github_graph_gophers_graphql_go//relay",
@@ -39,10 +39,10 @@ go_test(
         "//cmd/frontend/graphqlbackend",
         "//enterprise/cmd/frontend/internal/batches/resolvers/apitest",
         "//enterprise/cmd/frontend/internal/notebooks/resolvers/apitest",
-        "//enterprise/internal/notebooks",
         "//internal/actor",
         "//internal/database",
         "//internal/database/dbtest",
+        "//internal/notebooks",
         "//internal/types",
         "//lib/errors",
         "@com_github_google_go_cmp//cmp",
