+++ internal/notebooks/validate.go
